backend/requirements.txt
fastapi
uvicorn
pdfplumber
numpy
scikit-learn
requests
scipy
google-generativeai>=0.3.0
python-dotenv>=1.0.0
python-multipart
openai>=1.3.0

